[FIX] Fixed amount calculation from account move line
When a Ri.Ba. is created, the amount is calculated on amount_residual.
This field compute the amount without signs, mixing credit and debit.
So, when a list of credit and debit lines was selected for a riba issue,
the sum is not correct (it sums all debits and all credits).
The right sum indeed will be: debit sum minus credit sum.

Given that plugin already extends account_due_list and this define the
field maturity_residual (that compute correctly the total), I changed
the field used from amount_residual to maturity_residual.