[FIX] Invalid Account Move Line credit value when confirm a riba.
The move line created from the riba line use the amount fields instead
of using credit and debit fields of the related move line. This can
raise an issue when a credit note is added: the negative amount will
fail with the credit_debit2 constraint (credit+debit>0) because for this
move line the amount is negative.